+++ lib/dicey/cli/formatters/list_formatter.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+require_relative "key_value_formatter"
+
+module Dicey
+  module CLI
+    module Formatters
+      # Formats a hash as list of key => value pairs, similar to a Ruby Hash.
+      class ListFormatter < KeyValueFormatter
+        SEPARATOR = " => "
+      end
+    end
+  end
+end