+++ spec/dicey/roller_spec.rb
@@ -5,7 +5,7 @@ module Dicey
     subject(:result) { described_class.new.call(dice, format: format) }
 
     let(:dice) { %w[2 (1,5) 2D3] }
-    let(:format) { OutputFormatters::ListFormatter.new }
+    let(:format) { CLI::Formatters::ListFormatter.new }
 
     it "returns a formatted string with dice description and roll result" do
       expect(result).to match(/\A# D2\+\(1,5\)\+D3\+D3\nroll => (?:[4-9]|1[0-3])\n\z/)