+++ spec/dicey/sum_frequency_calculators/runner_spec.rb
@@ -6,7 +6,7 @@ module Dicey
 
     let(:dice) { %w[2d2 1,5,] } # rubocop:disable Lint/PercentStringArray
     let(:calculators) { [SumFrequencyCalculators::KroneckerSubstitution.new] }
-    let(:format) { OutputFormatters::JSONFormatter.new }
+    let(:format) { CLI::Formatters::JSONFormatter.new }
     let(:result) { :frequencies }
 
     before { stub_const("Dicey::SumFrequencyCalculators::AutoSelector::AVAILABLE_CALCULATORS", calculators) }